Handle quitting gracefully in the main class
Because we had threads initialized before the signal handler was installed
there was a race-condition if SIGTERM or SIGINT were sent during this time,
resulting in a crash.
The way we handle right now, there should be no chance of race condition
because we install the handlers before starting the threads.

The worst thing is that we have to wait to initialize everything
before we can actually quit.
