+++ homesensorhub/homesensorhub.py
@@ -1,5 +1,4 @@
 """Entry point for the application."""
-from routing.sensor_sink import SourceAndSink
 from routing.mqtt_sender import MQTTSender
 from routing.mqtt_subscriber import MQTTSubscriber
 
@@ -8,14 +7,18 @@ from sensors.BoschBME680.probe import ProbeBoschBME680
 from sensors.TSL258x.probe import ProbeTSL258x
 from sensors.RCWL0515.probe import ProbeRCWL0515
 
-from utils import logging
+from signal import signal, SIGINT, SIGTERM
+from threading import Event
 
+from utils import logging
 logger = logging.getLogger(__name__)
 
 
 class HomeSensorHub:
     def __init__(self):
         logger.success("Starting HomeSensorHub ...")
+        self.initialize_stop()  # initialize stop here so we are prepared
+
         self.__mqtt_subscriber = MQTTSubscriber()
 
         probe_functions = self.find_probe_functions()
@@ -24,11 +27,24 @@ class HomeSensorHub:
 
         self.__mqtt_subscriber.subscribe_to_sensor_properties(self.__sensors)
 
+    def initialize_stop(self):
+        self.event = Event()
+        signal(SIGINT, self.stop_everything)
+        signal(SIGTERM, self.stop_everything)
+
+    def stop_everything(self, signal_received, frame) -> None:
+        logger.notice("QUIT signal received !")
+        self.event.set()
+
     def run(self):
         logger.success("Running HomeSensorHub ...")
-        sensor_sink = SourceAndSink(self.__sensors, self.__senders)
-        sensor_sink.sink_and_send()
-        logger.success("Quitting gracefully")
+
+        # just wait for the stop event
+        self.event.wait()
+        for sensor in self.__sensors:
+            sensor.stop()
+
+        print("Quitting gracefully")
 
     def find_probe_functions(self):
         probe_functions = []
